requirements.txt
telethon==1.34.0
rich==13.7.0
python-dotenv==1.0.1
typer==0.9.0
